mensajeria-asincronica/02-java-rabbitmq-python/python/requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.1.8
flasgger==0.9.7.1
Flask==3.1.0
gunicorn==23.0.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==24.2
pika==1.3.2
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
typing_extensions==4.13.0
Werkzeug==3.1.3
zipp==3.21.0
